refactor(developments): extract developments code from listings context to its own
