+++ apps/re/lib/listings/schemas/listing.ex
@@ -123,9 +123,9 @@ defmodule Re.Listing do
     |> calculate_price_per_area()
   end
 
-  @development_required ~w(type description has_elevator address_id user_id development_uuid)a
+  @development_required ~w(type description has_elevator address_id development_uuid)a
 
-  @development_optional ~w(matterport_code is_exclusive status is_release is_exportable)a
+  @development_optional ~w(matterport_code is_exclusive status user_id is_release is_exportable)a
 
   @development_attributes @development_required ++ @development_optional
