+++ apps/re/lib/listings/units/propagator.ex
@@ -4,7 +4,7 @@ defmodule Re.Listings.Units.Propagator do
   be replicated/reflected in listings until we migrate replicated structure to units.
   """
 
-  alias Re.Listings
+  alias Re.Developments.Listings
 
   @cloned_attributes ~w(complement price property_tax maintenance_fee floor rooms bathrooms
     restrooms area garage_spots garage_type suites dependencies balconies)a
