+++ apps/re/test/listings/listings_test.exs
@@ -432,52 +432,6 @@ defmodule Re.ListingsTest do
                  owner_contact: nil
                )
     end
-
-    @insert_development_listing_params %{
-      "type" => "Apartamento",
-      "has_elevator" => true,
-      "description" => "Awesome new brand building",
-      "is_exportable" => true
-    }
-
-    test "should insert development listing" do
-      address = insert(:address)
-      development = insert(:development, address_id: address.id)
-      user = insert(:user, role: "admin")
-
-      assert {:ok, inserted_listing} =
-               Listings.insert(@insert_development_listing_params,
-                 address: address,
-                 user: user,
-                 development: development
-               )
-
-      assert retrieved_listing = Repo.get(Listing, inserted_listing.id)
-      assert retrieved_listing.development_uuid == development.uuid
-      assert retrieved_listing.address_id == address.id
-      assert retrieved_listing.user_id == user.id
-      assert retrieved_listing.is_exportable == true
-      assert retrieved_listing.uuid
-    end
-
-    test "should copy infrastructure info from development" do
-      address = insert(:address)
-      development = insert(:development, address_id: address.id)
-      user = insert(:user, role: "admin")
-
-      assert {:ok, inserted_listing} =
-               Listings.insert(@insert_development_listing_params,
-                 address: address,
-                 user: user,
-                 development: development
-               )
-
-      assert retrieved_listing = Repo.get(Listing, inserted_listing.id)
-      assert retrieved_listing.development_uuid == development.uuid
-      assert retrieved_listing.floor_count == development.floor_count
-      assert retrieved_listing.unit_per_floor == development.units_per_floor
-      assert retrieved_listing.elevators == development.elevators
-    end
   end
 
   describe "update/4" do
